show unhealthy agents first
should probably use the agent health details
